refactor: colocate admin core route tests
